+++ mlem/contrib/heroku/server.py
@@ -9,7 +9,7 @@ logger = logging.getLogger(__name__)
 
 
 class HerokuServer(FastAPIServer):
-    type: ClassVar = "heroku"
+    type: ClassVar = "_heroku"
 
     def serve(self, interface: Interface):
         self.port = int(os.environ["PORT"])
